+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Here is the list of the changes and features provided by the **node-red-contrib-ale-rainbow**
 
+
+## [1.3.4] - 2018-09-03
+- Login node: add appID and secret which will be mandatory soon in production
+- Send_IM node: allow to also modify Bubble's customData (only for Bubble owner)
+Thanks to loloj
+- Send_Channel and Notified_Channel nodes added
+- Add some icons for Rainbow Nodes...
+
 ## [1.3.3] - 2018-07-00
 - Add new node Rainbow_function to allow to call a specific API in Rainbow Node.js SDK
 Thanks to Christian Foricher